%% Documentation of lods2 data-frame used in examples
\name{lods2}
\docType{data}
\alias{lods2}
\title{LOD score table for chromosome 2}
\description{
This is a data frame with the first two columns containing marker names
and positions, followed by three columns of LOD scores.  
}
\usage{data(lods2)}
\format{
  There are 87 markers in the table. 
}
\keyword{datasets}
