\name{grid-internal}
%% These variables are imported from the example header files
%% lods1header.R and lods2header.R. These header files are meant to be
%% evaluated from within the nplplot.multi() function, otherwise they
%% have no other utility
\alias{yline}
\alias{ymin}
\alias{ymax}
\alias{yfix}
\alias{title}
\alias{ylabel}
\alias{draw.lgnd}
\alias{cex.legend}
\alias{cex.axis}
\alias{tcl}
\alias{bw}
\alias{na.rm}
\alias{lgndtxt}
\alias{ptypes}
\alias{my.colors}
\alias{ltypes}
\title{Internal nplplot.multi variables}
\description{
  Internal nplplot.multi
}
\details{
  These are not to be called by the user. These are added to the
  namespace when nplplot.multi() reads the lods1header.R and
  lods2header.R files.
}
\keyword{ internal }
