%% Documentation of lods2 data-frame used in examples
\name{lods2}
\docType{data}
\alias{lods2}
\title{LOD score table for chromosome 1}
\description{
This is a data-frame with the first two columns containing Marker names
and positions, followed by one or more columns of LOD-scores.  
}
\usage{lods2}
\format{
  There are 101 markers in the table.
}
\keyword{datasets}